Client#request(): Wait to recieve a socket to set the Date header.
http{s,}.Agent instances can introduce arbitrary delays between the creation of the request and the sending of the headers, which can cause the Date to be skewed according to Amazon's servers.

In particular, when making many requests in parallel with the default agent, if the first N-5 reqs take more than 15 minutes, then the Nth req will fail if the Date header is set immediately.

Once a socket is recieved, data (particularly, the request headers) can be sent immediately